Fix a glitch:
Let's say there are a chain of nodes: A-B-C-D.

The D is still able to connect to A if it wants to, even if D submited all the partner data correctly to A, A will still take the access request and grant it.

That's because node A don't know which partner is itself by looking those node IP addresses submitted by node D. It's Self-Address paradox: A host can't actually know all of the IP addresses can be used to access itself (If you consider there may proxies in the middle or maybe A in a LAN exposed port mapping etc).

I solved this problem by changed the node mutex mechanism to let node not connect with a node which already in the node path (In this case, let D don't connect with A voluntary), which is reduced the chance of this problem, but not fixed.

I also working on another method that scans connected node for confilct detection, and I plan to implement it with the connection construct functions so we can by the way build a faster network of connection.